move surface nets to jsm/libs
